+++ crates/factorio-exporter/src/internal/mod.rs
@@ -1,3 +1,1 @@
 pub(crate) mod mod_controller;
-pub(crate) mod script_builder;
-pub(crate) mod script_generator;